Add files via upload
This model is used to predict the gender and age of a person using Deep Learning Networks. The users are given 2 options. Option 1 being to upload a photo to predict the age and gender of the photo. option 2 being to activate the webcam so that the system monitors the age of the user who is in front of the web cam